requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
joblib==1.1.0
mathparse==0.1.2
nltk==3.6.5
numpy==1.21.4
packaging==21.3
pandas==1.3.5
Pint==0.18
psycopg2==2.9.2
pymongo==3.12.3
pyparsing==3.0.6
python-dateutil==1.5
pytz==2021.3
PyYAML==3.13
regex==2021.11.10
requests==2.26.0
six==1.16.0
SQLAlchemy==1.2.19
sqlparse==0.4.2
tqdm==4.62.3
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
